Move aggregate tests to a new file
For the upcoming tests, I need a third column in the 'foo' table -- or
alternatively a second table. Extending the existing table in
`test_integration.rb` let some other tests fail. Given that
`test_integration.rb` is already lengthy, I thought it would make
things clearer if I separate the aggregate tests into their own file
with their own setup/teardown methods.
